refactor: move CLI E2E tests into test/e2e/ directory
Separates the three test tiers by directory:
- commands/ → unit tests (mock engine/IO)
- e2e/     → E2E tests (spawn CLI binary)
- integration/ → integration tests (real engine, fixtures)
