+++ packages/cli/test/e2e/cli-from-to-dir.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import * as fs from 'fs/promises';
 import * as path from 'path';
-import { runCli, createTempDir, removeTempDir } from './test-support/cli-runner.js';
+import { runCli, createTempDir, removeTempDir } from '../test-support/cli-runner.js';
 
 describe('CLI --from-dir and --to-dir flags', () => {
   let tempDir: string;
